-Remove data layer from app module -Rename clases
