--- data/src/main/java/io/parrotsoftware/qa_data/UserManagerD.kt
@@ -1,17 +1,0 @@
-package io.parrotsoftware.qa_data
-
-interface UserManagerD {
-    fun saveCredentials(access: String, refresh: String)
-
-    fun getAccess(): String
-
-    fun getRefresh(): String
-
-    fun saveStore(uuid: String, name: String)
-
-    fun getStoreUuid(): String
-
-    fun getStoreName(): String
-
-    fun isAuth(): Boolean
-}
