+++ data/src/main/java/io/parrotsoftware/qa_data/repositories/UserRepository.kt
@@ -3,14 +3,14 @@ package io.parrotsoftware.qa_data.repositories
 import io.parrotsoftware.qa_data.domain.Credentials
 import io.parrotsoftware.qa_data.domain.RepositoryResult
 import io.parrotsoftware.qa_data.domain.Store
-import io.parrotsoftware.qa_data.UserManagerD
+import io.parrotsoftware.qa_data.managers.UserManager
 import io.parrotsoftware.qa_data.datasources.UserRemoteDataSource
 import javax.inject.Inject
 
 class UserRepository
 @Inject constructor(
     private val userRemoteDataSource: UserRemoteDataSource,
-    private val userManagerD: UserManagerD
+    private val userManagerD: UserManager
     ) {
 
     suspend fun userExists(): RepositoryResult<Boolean> =
